Rewrote grid to use <table />
https://github.com/SoftUni-Internal/exam-systems-issues/issues/1171
