+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/components/submissions/submissions-grid/SubmissionsGrid.tsx
@@ -5,7 +5,9 @@ import { IDictionary } from '../../../common/common-types';
 import { ISubmissionResponseModel } from '../../../common/types';
 import { usePublicSubmissions } from '../../../hooks/submissions/use-public-submissions';
 import { usePages } from '../../../hooks/use-pages';
+import useTheme from '../../../hooks/use-theme';
 import { IAuthorizationReduxState } from '../../../redux/features/authorizationSlice';
+import concatClassNames from '../../../utils/class-names';
 import { format } from '../../../utils/number-utils';
 import { flexCenterObjectStyles } from '../../../utils/object-utils';
 import Heading, { HeadingType } from '../../guidelines/headings/Heading';
@@ -44,6 +46,8 @@ const SubmissionsGrid = () => {
             clearPageInformation,
         },
     } = usePublicSubmissions();
+
+    const { isDarkMode, getColorClassName, themeColors } = useTheme();
 
     const { internalUser: user } =
         useSelector((state: {authorization: IAuthorizationReduxState}) => state.authorization);
@@ -165,6 +169,14 @@ const SubmissionsGrid = () => {
         [],
     );
 
+    const headerClassName = concatClassNames(
+        styles.submissionsGridHeader,
+        isDarkMode
+            ? styles.darkSubmissionsGridHeader
+            : styles.lightSubmissionsGridHeader,
+        getColorClassName(themeColors.textColor),
+    );
+
     const renderSubmissionsList = useCallback(
         () => {
             if (areSubmissionsLoading) {
@@ -184,18 +196,29 @@ const SubmissionsGrid = () => {
             }
 
             return (
-                <List
-                  values={publicSubmissions}
-                  itemFunc={renderSubmissionRow}
-                  fullWidth
-                />
+                <table className={styles.submissionsGrid}>
+                    <tr className={headerClassName}>
+                        <td>ID</td>
+                        <td>Task</td>
+                        <td>From</td>
+                        <td className={styles.tdRight}>Result</td>
+                        {
+                            user.isAdmin
+                                ? <td>Execution Result</td>
+                                : null
+                        }
+                        <td />
+                        <td />
+                    </tr>
+                    {
+                        publicSubmissions.map((s) => (
+                            <SubmissionGridRow submission={s} />
+                        ))
+                    }
+                </table>
             );
         },
-        [
-            publicSubmissions,
-            renderSubmissionRow,
-            areSubmissionsLoading,
-        ],
+        [ areSubmissionsLoading, publicSubmissions, user.isAdmin ],
     );
 
     return (
@@ -213,7 +236,7 @@ const SubmissionsGrid = () => {
             </Heading>
             {renderPrivilegedComponent()}
             {renderSubmissionsList()}
-            { publicSubmissions?.length > 0 && (
+            {publicSubmissions?.length > 0 && (
                 <PaginationControls
                   count={pagesCount}
                   page={currentPage}
